Use local paths for templates
Address #50

Fix the template paths as they were using the remote host paths since
previous testing of the scripts was only on localhost. Switched them
to use {{ inventory_dir }}.

Reorder when and tags in install.yaml to see if it matters as it
appeared that when was overriding tags when tags are spedified,
although I'm still confused about how this works. It is stupid to
think that the order of tags and when matters.